+++ src/main/java/com/pvmhud/overlay/SegmentKind.java
@@ -2,6 +2,7 @@ package com.pvmhud.overlay;
 
 enum SegmentKind {
     STAT,
+    POTION,
     SPELL,
     HEART
 }